Merged Writer and Format Writer into a single protocol.
This greatly simplifies the Logger -> Formatter -> Writer handoff process. It also allows us to remove the @objc annotations because we no longer need to downcast protocols. I also modified the initialization of the formatters and writers parameters in the Logger. Everything is now clearly defined so the consumer knows what the Logger will do by default.